Refactor LotteryController: Use Rails Naming Conventions and Extract Business Logic to Service Object
